Add files via upload
The csv files in the project are the files to be worked on. Each file contains numeric data stored in rows and columns.
The main.py file contains a python script used to loop through the csv files and then calculate the minimum, maximum and average of each column in each of the csv files.
The minimums, maximums and averages are then stored in three different csv files all at once.